Bump base image to newer Ubuntu (#353)
It's time! I don't actually think we need to wait for 20.4.1 - I think that
is valid for desktop systems and VMs, but not for container images. 
Particularly, we don't use the subsysems that are the primary
reason for 'wait for first point release' - the kernel, filesystems and the desktop
environment.

Fixes #352