Pin jaxlib to use cuda120 build with cuda-nvcc and update README note (#549)
* Pin jaxlib to use cuda120 build

The conda-lock solver was picking up the cpu build of jaxlib instead of the cuda build. Adding a explicit pin here to pick up the cuda120 build version.

* Update note on cuda-nvcc workaround

* Add unit test for jax random number generator

Regression test to ensure that JaX and cuda-nvcc are installed and compatible on GPU devices.

* Wrap jax test on GPU in try-except block to pass on GH Actions CI

Since we're using a version of GitHub Actions without a GPU, the JaX random number generator test on GPU cannot be properly tested, so wrapping the check in a try-except block. Also tidied up some import statements and docstrings in the test file.

* Update note in README to say cuda-nvcc workaround is no longer needed

Since `cuda-nvcc` is installed with jaxlib's cuda120 builds, the workaround to conda install cuda-nvcc should not be needed anymore.

* Add link to XLA (Accelerated Linear Algebra) page

Link to https://openxla.org/xla so that people know what XLA is.